extract common fixtures from inference and alteration tests
