+++ db/tests/types/test_inference.py
@@ -1,36 +1,16 @@
 import pytest
 from sqlalchemy import Column, MetaData, Table
 from sqlalchemy import BOOLEAN, Numeric, NUMERIC, String, VARCHAR
-from sqlalchemy.schema import CreateSchema, DropSchema
-from db.engine import _add_custom_types_to_engine
-from db.types import base, inference, install
-
-TEST_SCHEMA = "test_schema"
+from db.tests.types import fixtures
+from db.types import inference
 
 
-@pytest.fixture
-def engine_with_types(engine):
-    _add_custom_types_to_engine(engine)
-    return engine
-
-
-@pytest.fixture
-def temporary_testing_schema(engine_with_types):
-    schema = TEST_SCHEMA
-    with engine_with_types.begin() as conn:
-        conn.execute(CreateSchema(schema))
-    yield engine_with_types, schema
-    with engine_with_types.begin() as conn:
-        conn.execute(DropSchema(schema, cascade=True, if_exists=True))
-
-
-@pytest.fixture
-def engine_email_type(temporary_testing_schema):
-    engine, schema = temporary_testing_schema
-    install.install_mathesar_on_database(engine)
-    yield engine, schema
-    with engine.begin() as conn:
-        conn.execute(DropSchema(base.SCHEMA, cascade=True, if_exists=True))
+# We need to set these variables when the file loads, or pytest can't
+# properly detect the fixtures.  Importing them directly results in a
+# flake8 unused import error.
+engine_with_types = fixtures.engine_with_types
+engine_email_type = fixtures.engine_email_type
+temporary_testing_schema = fixtures.temporary_testing_schema
 
 
 type_data_list = [
